Fixes to get picker working in Umbraco 16
Implements a configuration editor and property value editor.

Updates the UI package to include a property editor schema.

Uses the EditorAlias not DataEditorAlias in the property value converter so the converter is actually called (possibly not necessary now there is a named data editor instead of the default JSON one, but this is consistent with core property value converters).